Added upgrade test
