+++ test/e2e/suite/cases/doc.go
@@ -5,4 +5,5 @@ import (
 	_ "github.com/jetstack/kube-oidc-proxy/test/e2e/suite/cases/passthrough"
 	_ "github.com/jetstack/kube-oidc-proxy/test/e2e/suite/cases/rbac"
 	_ "github.com/jetstack/kube-oidc-proxy/test/e2e/suite/cases/token"
+	_ "github.com/jetstack/kube-oidc-proxy/test/e2e/suite/cases/upgrade"
 )
