+++ test/e2e/suite/cases/passthrough/passthrough.go
@@ -1,4 +1,4 @@
-package impersonation
+package passthrough
 
 import (
 	"fmt"
